finalized spendingMoney() functionality
