+++ src/javascripts/helpers/data/visitorData/visitorsData.js
@@ -1,8 +1,5 @@
 import axios from 'axios';
 import firebaseConfig from '../../auth/apiKeys';
-import addLog from '../logData';
-import { getRides } from '../rideData/ridesData';
-import { getVendors } from '../vendorData/vendorData';
 
 const dbUrl = firebaseConfig.databaseURL;
 
@@ -46,53 +43,6 @@ const updateVisitor = (firebaseKey, visitorObject) => new Promise((resolve, reje
     .catch((error) => reject(error));
 });
 
-// MASTER RIDES AND VENDORS LIST
-const ridesAndVendors = () => new Promise((resolve, reject) => {
-  Promise.all([getRides(), getVendors()])
-    .then(([rides, vendors]) => {
-      // BUILD ARRAY OF RIDES AND VENDORS TO CHOOSE FROM
-      const ridesAndVendorsArray = [];
-      rides.forEach((ride) => {
-        const object = {
-          earnings: ride.price,
-          event: ride.rideName
-        };
-        ridesAndVendorsArray.push(object);
-      });
-      vendors.forEach((vendor) => {
-        const object = {
-          earnings: vendor.price,
-          event: vendor.vendorName
-        };
-        ridesAndVendorsArray.push(object);
-      });
-      resolve(ridesAndVendorsArray);
-    })
-    .catch((error) => reject(error));
-});
-
-// SPENDING MONEY
-const spendingMoney = () => {
-  ridesAndVendors().then((resultsArray) => {
-    // (B) GET ALL VISITORS
-    const logArray = [];
-    getVisitors().then((visitorsArray) => {
-    // START BUILDING LOG
-      visitorsArray.forEach((visitor) => {
-        const randomEvent = resultsArray[Math.floor(Math.random() * resultsArray.length)];
-        const object = {
-          visitor: visitor.visitorFirstName,
-          ...randomEvent,
-          timestamp: new Date()
-        };
-        logArray.push(object);
-      });
-      // (C) POST LOG TO FIREBASE
-      addLog(logArray);
-    });
-  });
-};
-
 export {
-  getVisitors, addVisitor, deleteVisitor, updateVisitor, getSingleVisitor, spendingMoney
+  getVisitors, addVisitor, deleteVisitor, updateVisitor, getSingleVisitor
 };
